refactor(gui): split GenealogyTimelineView by responsibility
